requirements.txt
azure-cognitiveservices-speech == 1.37.0
azure-storage-blob == 12.19.1